+++ opm/grid/CellQuadrature.cpp
@@ -68,8 +68,8 @@ inline double triangleArea2d(const double* p0,
 namespace Opm {
 
 CellQuadrature::CellQuadrature(const UnstructuredGrid& grid,
-                               const int cell,
-                               const int degree)
+                               const long long cell,
+                               const long long degree)
     : grid_(grid), cell_(cell), degree_(degree)
 {
     if (grid.dimensions > 3) {
@@ -80,7 +80,7 @@ CellQuadrature::CellQuadrature(const UnstructuredGrid& grid,
     }
 }
 
-int CellQuadrature::numQuadPts() const
+long long CellQuadrature::numQuadPts() const
 {
     if (degree_ < 2 || grid_.dimensions == 1) {
         return 1;
@@ -90,17 +90,17 @@ int CellQuadrature::numQuadPts() const
         return 3*(grid_.cell_facepos[cell_ + 1] - grid_.cell_facepos[cell_]);
     }
     assert(grid_.dimensions == 3);
-    int sumnodes = 0;
+    long long sumnodes = 0;
     for (unsigned hf = grid_.cell_facepos[cell_]; hf < grid_.cell_facepos[cell_ + 1]; ++hf) {
-        const int face = grid_.cell_faces[hf];
+        const long long face = grid_.cell_faces[hf];
         sumnodes += grid_.face_nodepos[face + 1] - grid_.face_nodepos[face];
     }
     return 4*sumnodes;
 }
 
-void CellQuadrature::quadPtCoord(const int index, double* coord) const
+void CellQuadrature::quadPtCoord(const long long index, double* coord) const
 {
-    const int dim = grid_.dimensions;
+    const long long dim = grid_.dimensions;
     const double* cc = grid_.cell_centroids + dim*cell_;
     if (degree_ < 2) {
         std::copy(cc, cc + dim, coord);
@@ -110,48 +110,48 @@ void CellQuadrature::quadPtCoord(const int index, double* coord) const
     if (dim == 2) {
         if (index % 3 == 0) {
             // Boundary midpoint. This is the face centroid.
-            const int hface = grid_.cell_facepos[cell_] + index/3;
-            const int face = grid_.cell_faces[hface];
+            const long long hface = grid_.cell_facepos[cell_] + index/3;
+            const long long face = grid_.cell_faces[hface];
             const double* fc = grid_.face_centroids + dim*face;
             std::copy(fc, fc + dim, coord);
         } else {
             // Interiour midpoint. This is the average of the
             // cell centroid and a face node (they should
             // always have two nodes in 2d).
-            const int hface = grid_.cell_facepos[cell_] + index/3;
-            const int face = grid_.cell_faces[hface];
-            const int nodeoff = (index % 3) - 1; // == 0 or 1
-            const int node = grid_.face_nodes[grid_.face_nodepos[face] + nodeoff];
+            const long long hface = grid_.cell_facepos[cell_] + index/3;
+            const long long face = grid_.cell_faces[hface];
+            const long long nodeoff = (index % 3) - 1; // == 0 or 1
+            const long long node = grid_.face_nodes[grid_.face_nodepos[face] + nodeoff];
             const double* nc = grid_.node_coordinates + dim*node;
-            for (int dd = 0; dd < dim; ++dd) {
+            for (long long dd = 0; dd < dim; ++dd) {
                 coord[dd] = 0.5*(nc[dd] + cc[dd]);
             }
         }
         return;
     }
     assert(dim == 3);
-    int tetindex = index / 4;
-    const int subindex = index % 4;
+    long long tetindex = index / 4;
+    const long long subindex = index % 4;
     const double* nc = grid_.node_coordinates;
     for (unsigned hf = grid_.cell_facepos[cell_]; hf < grid_.cell_facepos[cell_ + 1]; ++hf) {
-        const int face = grid_.cell_faces[hf];
-        const int nfn = grid_.face_nodepos[face + 1] - grid_.face_nodepos[face];
+        const long long face = grid_.cell_faces[hf];
+        const long long nfn = grid_.face_nodepos[face + 1] - grid_.face_nodepos[face];
         if (nfn <= tetindex) {
             // Our tet is not associated with this face.
             tetindex -= nfn;
             continue;
         }
         const double* fc = grid_.face_centroids + dim*face;
-        const int* fnodes = grid_.face_nodes + grid_.face_nodepos[face];
-        const int node0 = fnodes[tetindex];
-        const int node1 = fnodes[(tetindex + 1) % nfn];
+        const long long* fnodes = grid_.face_nodes + grid_.face_nodepos[face];
+        const long long node0 = fnodes[tetindex];
+        const long long node1 = fnodes[(tetindex + 1) % nfn];
         const double* n0c = nc + dim*node0;
         const double* n1c = nc + dim*node1;
         const double a = 0.138196601125010515179541316563436;
         // Barycentric coordinates of our point in the tet.
         double baryc[4] = { a, a, a, a };
         baryc[subindex] = 1.0 - 3.0*a;
-        for (int dd = 0; dd < dim; ++dd) {
+        for (long long dd = 0; dd < dim; ++dd) {
             coord[dd] = baryc[0]*cc[dd] + baryc[1]*fc[dd] + baryc[2]*n0c[dd] + baryc[3]*n1c[dd];
         }
         return;
@@ -159,37 +159,37 @@ void CellQuadrature::quadPtCoord(const int index, double* coord) const
     OPM_THROW(std::runtime_error, "Should never reach this point.");
 }
 
-double CellQuadrature::quadPtWeight(const int index) const
+double CellQuadrature::quadPtWeight(const long long index) const
 {
     if (degree_ < 2) {
         return grid_.cell_volumes[cell_];
     }
     // Degree 2 case.
-    const int dim = grid_.dimensions;
+    const long long dim = grid_.dimensions;
     const double* cc = grid_.cell_centroids + dim*cell_;
     if (dim == 2) {
-        const int hface = grid_.cell_facepos[cell_] + index/3;
-        const int face = grid_.cell_faces[hface];
-        const int* nptr = grid_.face_nodes + grid_.face_nodepos[face];
+        const long long hface = grid_.cell_facepos[cell_] + index/3;
+        const long long face = grid_.cell_faces[hface];
+        const long long* nptr = grid_.face_nodes + grid_.face_nodepos[face];
         const double* nc0 = grid_.node_coordinates + dim*nptr[0];
         const double* nc1 = grid_.node_coordinates + dim*nptr[1];
         return triangleArea2d(nc0, nc1, cc)/3.0;
     }
     assert(dim == 3);
-    int tetindex = index / 4;
+    long long tetindex = index / 4;
     const double* nc = grid_.node_coordinates;
     for (unsigned hf = grid_.cell_facepos[cell_]; hf < grid_.cell_facepos[cell_ + 1]; ++hf) {
-        const int face = grid_.cell_faces[hf];
-        const int nfn = grid_.face_nodepos[face + 1] - grid_.face_nodepos[face];
+        const long long face = grid_.cell_faces[hf];
+        const long long nfn = grid_.face_nodepos[face + 1] - grid_.face_nodepos[face];
         if (nfn <= tetindex) {
             // Our tet is not associated with this face.
             tetindex -= nfn;
             continue;
         }
         const double* fc = grid_.face_centroids + dim*face;
-        const int* fnodes = grid_.face_nodes + grid_.face_nodepos[face];
-        const int node0 = fnodes[tetindex];
-        const int node1 = fnodes[(tetindex + 1) % nfn];
+        const long long* fnodes = grid_.face_nodes + grid_.face_nodepos[face];
+        const long long node0 = fnodes[tetindex];
+        const long long node1 = fnodes[(tetindex + 1) % nfn];
         const double* n0c = nc + dim*node0;
         const double* n1c = nc + dim*node1;
         return 0.25*tetVolume(cc, fc, n0c, n1c);
